Era por ela asignacion del Byte

+++ Persistencia/Domain/Alojamiento.cs
@@ -24,7 +24,7 @@ namespace Persistencia.Domain
 
         //-------------------- Relaciones ---------------------------------//
     //    public int ClienteId { get; set; }
-        public int HabitacionId { get; set; }
+        public byte HabitacionId { get; set; }
         public Habitacion Habitacion { get; set; }
         public List<Cliente> Clientes { get; set; }
         public List<Pago> Pagos { get; set; }
